backend/requirements.txt
flask>=3.0
flask-cors>=4.0
psycopg2-binary>=2.9
PyJWT>=2.8
bcrypt>=4.1
requests>=2.31
gunicorn>=21.2
